Bug 1328337 - fix implicit conversion of glyph size from float to uint32_t in NativeFontResource::CreateScaledFont. r=bobowen
MozReview-Commit-ID: 9JcrfRYJ3ca

UltraBlame original commit: aea479acaeb8ed3ac0f48235ff349042b82a61ba

+++ gfx/2d/2D.h
@@ -741,7 +741,7 @@ public:
 
 
   virtual already_AddRefed<ScaledFont>
-    CreateScaledFont(uint32_t aIndex, uint32_t aGlyphSize) = 0;
+    CreateScaledFont(uint32_t aIndex, Float aGlyphSize) = 0;
 
   virtual ~NativeFontResource() {};
 };
